Finish JSOM import builder refactoring
